修改綱要

+++ database/migrations/2024_10_22_040324_create_survey_responses_table.php
@@ -18,11 +18,11 @@ class CreateSurveyResponsesTable extends Migration
             $table->string('category', 50)->comment('統計項目別'); // 統計項目別
             $table->string('subcategory', 50)->comment('細項'); // 細項
             $table->integer('sample_size')->comment('樣本數（人）'); // 樣本數（人）
-            $table->decimal('very_satisfied_pct', 5, 2)->comment('很滿意（%）'); // 很滿意（%）
-            $table->decimal('satisfied_pct', 5, 2)->comment('滿意（%）'); // 滿意（%）
-            $table->decimal('neutral_pct', 5, 2)->comment('普通（%）'); // 普通（%）
-            $table->decimal('dissatisfied_pct', 5, 2)->comment('不滿意（%）'); // 不滿意（%）
-            $table->decimal('very_dissatisfied_pct', 5, 2)->comment('很不滿意（%）'); // 很不滿意（%）
+            $table->double('very_satisfied_pct')->comment('很滿意（%）'); // 很滿意（%）
+            $table->double('satisfied_pct')->comment('滿意（%）'); // 滿意（%）
+            $table->double('neutral_pct')->comment('普通（%）'); // 普通（%）
+            $table->double('dissatisfied_pct')->comment('不滿意（%）'); // 不滿意（%）
+            $table->double('very_dissatisfied_pct')->comment('很不滿意（%）'); // 很不滿意（%）
             $table->timestamps(); // 自動管理 created_at 和 updated_at
         });
     }
